Refactor add_impulse to use runtime booleans with const generics
- Refactor add_impulse to use const generics
- Extract quantization into a separate module
- Reduce visibility and reorganize linear_math
- Reorganize imports across multiple files
- Switch add_impulse to runtime booleans
