+++ rocketsim/src/bullet/collision/shapes/mod.rs
@@ -9,8 +9,8 @@ mod optimized_bvh;
 mod polyhedral_convex_shape;
 pub mod sphere_shape;
 pub mod static_plane_shape;
-pub mod triangle_callback;
-pub mod triangle_info_map;
+pub(super) mod triangle_callback;
+pub(super) mod triangle_info_map;
 pub mod triangle_mesh;
 mod triangle_mesh_shape;
-pub mod triangle_shape;
+pub(super) mod triangle_shape;
