+++ rocketsim/src/bullet/dynamics/constraint_solver/mod.rs
@@ -1,6 +1,6 @@
-pub mod contact_constraint;
-pub mod contact_solver_info;
+pub(super) mod contact_constraint;
+pub(super) mod contact_solver_info;
 mod jacobian_entry;
-pub mod seq_impulse_constraint_solver;
+pub(super) mod seq_impulse_constraint_solver;
 pub mod solver_body;
 mod solver_constraint;
