+++ rocketsim/src/bullet/dynamics/mod.rs
@@ -1,4 +1,4 @@
-pub mod constraint_solver;
+pub(super) mod constraint_solver;
 pub mod discrete_dynamics_world;
 pub mod rigid_body;
 pub mod vehicle;
