+++ rocketsim/src/shared/mod.rs
@@ -1,8 +1,11 @@
 mod aabb;
+mod angle;
 pub(crate) mod bvh;
 mod quad_ray;
+pub(crate) mod quantize;
 pub mod rand;
 pub mod rsmath;
 
 pub use aabb::*;
+pub use angle::Angle;
 pub use quad_ray::*;
